chore(lockfile): update package-lock.json to include react-icons dependency
This update adds the react-icons package version 5.5.0 to the project’s
package-lock.json file. The inclusion of this dependency allows for
the use of a comprehensive set of icons in the frontend application.
There are no known risks or side effects associated with this change,
as it is a standard dependency addition.